pyproject.toml
[project]
name = "automaton"
version = "0.1.0"
description = ""
authors = [
    {name = "trop3n",email = "jasondkimm@proton.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyvimeo (>=1.1.2,<2.0.0)",
    "datetime (>=5.5,<6.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "automaton", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
